Added some classes for defining Google Analytics events
