+++ frontend/src/analytics/AnalyticsEvent.ts
@@ -1,9 +1,11 @@
+// Defines the common enum to use for event names across all analytics services
 export enum AnalyticsEventEnum {
     BOOKING,
-    CANCEL_BOOKING
+    END_BOOKING
 }
 
+// Defines the common event names to use across all analytics services
 export const analyticsEventMap: { [key in AnalyticsEventEnum]: string } = {
     [AnalyticsEventEnum.BOOKING]: 'booking',
-    [AnalyticsEventEnum.CANCEL_BOOKING]: 'cancelBooking'
+    [AnalyticsEventEnum.END_BOOKING]: 'endBooking'
 };
