+++ frontend/src/components/AvailableRoomList/AvailableRoomList.tsx
@@ -17,7 +17,8 @@ import dayjs from 'dayjs';
 import { sortByFavoritedAndName } from '../../util/arrayUtils';
 import { triggerClarityEvent } from '../../analytics/clarityService';
 import { AnalyticsEventEnum } from '../../analytics/AnalyticsEvent';
-import { triggerGoogleAnalyticsEvent } from '../../analytics/googleAnalyticsService';
+import { triggerGoogleAnalyticsEvent } from '../../analytics/googleAnalytics/googleAnalyticsService';
+import { BookingEvent } from '../../analytics/googleAnalytics/googleAnalyticsEvents';
 
 const SKIP_CONFIRMATION = true;
 
@@ -291,7 +292,8 @@ const AvailableRoomList = (props: BookingListProps) => {
                 createSuccessNotification('Booking was successful');
                 setBookingLoading('false');
                 document.getElementById('main-view-content')?.scrollTo(0, 0);
-                triggerGoogleAnalyticsEvent(AnalyticsEventEnum.BOOKING);
+
+                triggerGoogleAnalyticsEvent(new BookingEvent(room));
                 triggerClarityEvent(AnalyticsEventEnum.BOOKING);
             })
             .catch(() => {
